chore: rename theme folders
