+++ src/templates/workspace/snippets/treeItemFile.ts
@@ -1,4 +1,4 @@
-import { getLangIcon } from '../../../theme2/getLangIcon'
+import { getLangIcon } from '../../../theme/getLangIcon'
 import { File, WorkspaceState } from '../../../webviews/Workspace/WorkspaceViewProvider.interface'
 import { RenderVars } from '../../../webviews/webviews.interface'
 import { getFileTooltip } from '../../helpers/getFileTooltip'
