+++ src/themeNpm/utils/fs/isHiddenFile.ts
