+++ src/themeNpm/utils/strings/cleanFileIconKey.ts
